Refactor : 카테고리별 조회 서비스에 기타 추가,String 길이 조정, limitAttendance 타입 short로 변경

+++ src/main/java/com/sparta/eroomprojectbe/domain/challenge/dto/ChallengeRequestDto.java
@@ -13,7 +13,7 @@ public class ChallengeRequestDto {
     private String startDate;
     private String dueDate;
     private String frequency;
-    private int limitAttendance;
+    private short limitAttendance;
     private String authExplanation;
     private String thumbnailImageUrl;
